Refactor : types 내의 TextType.ts 파일을 Text폴더의 내부로 옮김
