+++ src/components/text/textType.ts
